pages everywhere
## Separate the WebUI multiple pages:
- Home
- Firmwares
- Statistics
- Upload

## Created a modular dynamic component rendering system

Dynamicly injectable modals, javascript libs by pageWrapper
