+++ eslint.config.js
@@ -140,6 +140,16 @@ module.exports = [
                 bootstrap: "readonly",
                 fetch: "readonly",
                 FormData: "readonly",
+                showAlert: "readonly",
+                NetworkManager: "readonly",
+                AuthUpdateNetworkManager: "readonly",
+                compareSemver: "readonly",
+                getVersionBadgeClass: "readonly",
+                formatDate: "readonly",
+                formatFileSize: "readonly",
+                authToken: "readonly",
+                updateAuthUI: "readonly",
+                Chart: "readonly",
 			},
 		},
 	},
