+++ public/css/common.css
@@ -1,0 +1,88 @@
+.firmware-card {
+    transition: transform 0.2s, box-shadow 0.2s;
+}
+
+.firmware-card:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 4px 8px rgba(0,0,0,0.1);
+}
+
+.version-badge {
+    font-family: 'Courier New', monospace;
+}
+
+.upload-section {
+    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+    color: white;
+}
+
+.sha1-hash {
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.sha1-hash:hover {
+    background-color: rgba(0,0,0,0.1);
+    border-radius: 3px;
+    padding: 2px 4px;
+}
+
+.stats-card .row > div {
+    border-right: 1px solid #dee2e6;
+}
+
+.stats-card .row > div:last-child {
+    border-right: none;
+}
+
+.toast {
+    opacity: 0;
+    transition: opacity 0.3s ease;
+}
+
+.toast.show {
+    opacity: 1;
+}
+
+.server-status-online {
+    color: #28a745;
+}
+
+.server-status-offline {
+    color: #dc3545;
+}
+
+.btn-group .btn {
+    border-radius: 0;
+}
+
+.btn-group .btn:first-child {
+    border-top-left-radius: 0.375rem;
+    border-bottom-left-radius: 0.375rem;
+}
+
+.btn-group .btn:last-child {
+    border-top-right-radius: 0.375rem;
+    border-bottom-right-radius: 0.375rem;
+}
+
+.hero-section {
+    background: linear-gradient(135deg, #f8f9fa 0%, #e9ecef 100%);
+    border-radius: 1rem;
+    margin: 2rem 0;
+}
+
+.navbar-nav .nav-link.active {
+    font-weight: 600;
+    color: #fff !important;
+}
+
+.card {
+    border: none;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+    transition: box-shadow 0.2s;
+}
+
+.card:hover {
+    box-shadow: 0 4px 8px rgba(0,0,0,0.15);
+}
