+++ public/js/home.js
@@ -1,0 +1,4 @@
+document.addEventListener("DOMContentLoaded", () => {
+	// Home page specific functionality
+	console.log("Home page loaded");
+});
