--- public/js/main.js
@@ -1,962 +1,0 @@
-// Updated main.js with new features and endpoints
-
-let authToken = null;
-let currentUser = null;
-let firmwareStats = {};
-let currentView = "list"; // 'grid' or 'list'
-
-// Check for stored auth on page load
-if (typeof Storage !== "undefined") {
-	authToken = localStorage.getItem("authToken");
-	currentUser = localStorage.getItem("currentUser");
-}
-
-// Initialize
-document.addEventListener("DOMContentLoaded", () => {
-	updateAuthUI();
-	loadDevices();
-	loadFirmwares();
-	loadFirmwareStats();
-	createToastContainer();
-	setupEventListeners();
-	checkServerHealth();
-});
-
-function setupEventListeners() {
-	document.getElementById("deviceFilter").addEventListener("change", loadFirmwares);
-	document.getElementById("uploadForm").addEventListener("submit", uploadFirmware);
-
-	// Add refresh button listener
-	const refreshBtn = document.getElementById("refreshBtn");
-	if (refreshBtn) {
-		refreshBtn.addEventListener("click", () => {
-			refreshBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i>';
-			Promise.all([loadDevices(), loadFirmwares(), loadFirmwareStats(), checkServerHealth()]).finally(() => {
-				refreshBtn.innerHTML = '<i class="fas fa-sync-alt"></i>';
-			});
-		});
-	}
-
-	// File input validation
-	const fileInput = document.getElementById("firmwareFile");
-	if (fileInput) {
-		fileInput.addEventListener("change", validateFileInput);
-	}
-
-	// Keyboard shortcuts
-	document.addEventListener("keydown", handleKeyboardShortcuts);
-
-	// Auto-refresh every 5 minutes
-	setInterval(() => {
-		if (document.visibilityState === "visible") {
-			checkServerHealth();
-			loadFirmwareStats();
-		}
-	}, 5 * 60 * 1000);
-}
-
-function handleKeyboardShortcuts(event) {
-	// Ctrl/Cmd + R: Refresh
-	if ((event.ctrlKey || event.metaKey) && event.key === "r") {
-		event.preventDefault();
-		document.getElementById("refreshBtn").click();
-	}
-
-	// Ctrl/Cmd + U: Focus upload (if logged in)
-	if ((event.ctrlKey || event.metaKey) && event.key === "u" && authToken) {
-		event.preventDefault();
-		const uploadSection = document.getElementById("uploadSection");
-		if (!uploadSection.classList.contains("d-none")) {
-			document.getElementById("deviceType").focus();
-		}
-	}
-
-	// Escape: Close modals
-	if (event.key === "Escape") {
-		const openModals = document.querySelectorAll(".modal.show");
-		openModals.forEach((modal) => {
-			const modalInstance = bootstrap.Modal.getInstance(modal);
-			if (modalInstance) {
-				modalInstance.hide();
-			}
-		});
-	}
-
-	// Ctrl/Cmd + L: Focus device filter
-	if ((event.ctrlKey || event.metaKey) && event.key === "l") {
-		event.preventDefault();
-		document.getElementById("deviceFilter").focus();
-	}
-}
-
-function validateFileInput(event) {
-	const file = event.target.files[0];
-	if (!file) return;
-
-	const maxSize = 50 * 1024 * 1024; // 50MB
-	const allowedTypes = [".bin", ".hex", ".elf", ".ino", ".cpp", ".c", ".h"];
-
-	if (file.size > maxSize) {
-		showAlert("File size must be less than 50MB", "warning");
-		event.target.value = "";
-		return;
-	}
-
-	const fileExtension = `.${file.name.split(".").pop().toLowerCase()}`;
-	if (!allowedTypes.includes(fileExtension)) {
-		showAlert(`File type not supported. Please use: ${allowedTypes.join(", ")}`, "warning");
-		event.target.value = "";
-		return;
-	}
-}
-
-function toggleView() {
-	currentView = currentView === "grid" ? "list" : "grid";
-	const toggleBtn = document.getElementById("viewToggle");
-
-	if (currentView === "grid") {
-		toggleBtn.innerHTML = '<i class="fas fa-th-large"></i>';
-		toggleBtn.title = "Switch to List View";
-	} else {
-		toggleBtn.innerHTML = '<i class="fas fa-list"></i>';
-		toggleBtn.title = "Switch to Grid View";
-	}
-
-	// Reload firmwares with new view
-	loadFirmwares();
-}
-
-function createToastContainer() {
-	if (!document.getElementById("toastContainer")) {
-		const toastContainer = document.createElement("div");
-		toastContainer.id = "toastContainer";
-		toastContainer.style.cssText = `
-            position: fixed;
-            bottom: 20px;
-            right: 20px;
-            z-index: 9999;
-            max-width: 350px;
-        `;
-		document.body.appendChild(toastContainer);
-	}
-}
-
-function updateAuthUI() {
-	const loginBtn = document.getElementById("loginBtn");
-	const logoutBtn = document.getElementById("logoutBtn");
-	const uploadSection = document.getElementById("uploadSection");
-
-	if (authToken) {
-		loginBtn.classList.add("d-none");
-		logoutBtn.classList.remove("d-none");
-		logoutBtn.innerHTML = `<i class="fas fa-sign-out-alt me-1"></i>Logout (${currentUser})`;
-		uploadSection.classList.remove("d-none");
-
-		// Show edit and delete buttons
-		const editButtons = document.querySelectorAll('.btn-outline-warning[title="Edit"]');
-		const deleteButtons = document.querySelectorAll('.btn-outline-danger[title="Delete"]');
-
-		editButtons.forEach((btn) => (btn.style.display = "inline-block"));
-		deleteButtons.forEach((btn) => (btn.style.display = "inline-block"));
-	} else {
-		loginBtn.classList.remove("d-none");
-		logoutBtn.classList.add("d-none");
-		uploadSection.classList.add("d-none");
-
-		// Hide edit and delete buttons
-		const editButtons = document.querySelectorAll('.btn-outline-warning[title="Edit"]');
-		const deleteButtons = document.querySelectorAll('.btn-outline-danger[title="Delete"]');
-
-		editButtons.forEach((btn) => (btn.style.display = "none"));
-		deleteButtons.forEach((btn) => (btn.style.display = "none"));
-	}
-}
-
-function toggleAuth() {
-	if (authToken) {
-		logout();
-	} else {
-		new bootstrap.Modal(document.getElementById("loginModal")).show();
-	}
-}
-
-async function login() {
-	const username = document.getElementById("username").value;
-	const password = document.getElementById("password").value;
-
-	try {
-		const response = await fetch("/api/login", {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({ username, password }),
-		});
-
-		const data = await response.json();
-
-		if (response.ok) {
-			authToken = data.token;
-			currentUser = data.user?.username || "Admin";
-
-			if (typeof Storage !== "undefined") {
-				localStorage.setItem("authToken", authToken);
-				localStorage.setItem("currentUser", currentUser);
-			}
-
-			updateAuthUI();
-			bootstrap.Modal.getInstance(document.getElementById("loginModal")).hide();
-			document.getElementById("loginForm").reset();
-			showAlert("Login successful!", "success");
-		} else {
-			showAlert(data.error || "Login failed", "danger");
-		}
-	} catch (error) {
-		console.error("Login error:", error);
-		showAlert("Login failed - server error", "danger");
-	}
-}
-
-function logout() {
-	authToken = null;
-	currentUser = null;
-
-	if (typeof Storage !== "undefined") {
-		localStorage.removeItem("authToken");
-		localStorage.removeItem("currentUser");
-	}
-
-	updateAuthUI();
-	showAlert("Logged out successfully", "info");
-}
-
-async function checkServerHealth() {
-	try {
-		const response = await fetch("/api/health");
-		const health = await response.json();
-
-		if (response.ok && health.status === "healthy") {
-			console.log("Server health check passed");
-			displayServerStatus(true, health);
-		} else {
-			console.warn("Server health check failed", health);
-			displayServerStatus(false, health);
-		}
-	} catch (error) {
-		console.error("Health check error:", error);
-		displayServerStatus(false, { error: "Connection failed" });
-	}
-}
-
-function displayServerStatus(isHealthy, healthData) {
-	// Add a small status indicator to the navbar
-	const navbar = document.querySelector(".navbar .container");
-	let statusIndicator = document.getElementById("serverStatus");
-
-	if (!statusIndicator) {
-		statusIndicator = document.createElement("div");
-		statusIndicator.id = "serverStatus";
-		statusIndicator.className = "me-3";
-		navbar.insertBefore(statusIndicator, navbar.querySelector(".navbar-nav"));
-	}
-
-	const statusColor = isHealthy ? "success" : "danger";
-	const statusIcon = isHealthy ? "check-circle" : "exclamation-triangle";
-	const statusText = isHealthy ? "Online" : "Offline";
-
-	statusIndicator.innerHTML = `
-        <small class="text-${statusColor}">
-            <i class="fas fa-${statusIcon} me-1"></i>${statusText}
-        </small>
-    `;
-
-	// Also update footer status
-	const footerStatus = document.getElementById("footerServerStatus");
-	if (footerStatus) {
-		footerStatus.innerHTML = `
-            <small class="text-${statusColor}">
-                <i class="fas fa-${statusIcon} me-1"></i>${statusText}
-            </small>
-        `;
-	}
-}
-
-async function loadFirmwareStats() {
-	try {
-		const response = await fetch("/api/firmwares/stats");
-		const stats = await response.json();
-
-		if (response.ok) {
-			firmwareStats = stats;
-			displayFirmwareStats(stats);
-		}
-	} catch (error) {
-		console.error("Error loading firmware stats:", error);
-	}
-}
-
-function displayFirmwareStats(stats) {
-	// Add stats card before the device filter
-	const deviceFilterCard = document.querySelector(".card.mb-4");
-	let statsCard = document.getElementById("statsCard");
-
-	if (!statsCard) {
-		statsCard = document.createElement("div");
-		statsCard.id = "statsCard";
-		statsCard.className = "card mb-4";
-		deviceFilterCard.parentNode.insertBefore(statsCard, deviceFilterCard);
-	}
-
-	statsCard.innerHTML = `
-        <div class="card-body">
-            <div class="row text-center">
-                <div class="col-md-3">
-                    <div class="p-3">
-                        <i class="fas fa-microchip fa-2x text-primary mb-2"></i>
-                        <h4 class="mb-0">${stats.totalFirmwares || 0}</h4>
-                        <small class="text-muted">Total Firmwares</small>
-                    </div>
-                </div>
-                <div class="col-md-3">
-                    <div class="p-3">
-                        <i class="fas fa-layer-group fa-2x text-success mb-2"></i>
-                        <h4 class="mb-0">${stats.deviceTypes || 0}</h4>
-                        <small class="text-muted">Device Types</small>
-                    </div>
-                </div>
-                <div class="col-md-3">
-                    <div class="p-3">
-                        <i class="fas fa-hdd fa-2x text-info mb-2"></i>
-                        <h4 class="mb-0">${formatFileSize(stats.totalSize || 0)}</h4>
-                        <small class="text-muted">Total Size</small>
-                    </div>
-                </div>
-                <div class="col-md-3">
-                    <div class="p-3">
-                        <i class="fas fa-clock fa-2x text-warning mb-2"></i>
-                        <h4 class="mb-0">${stats.latestUpload ? formatDate(stats.latestUpload) : "N/A"}</h4>
-                        <small class="text-muted">Latest Upload</small>
-                    </div>
-                </div>
-            </div>
-        </div>
-    `;
-}
-
-async function loadDevices() {
-	try {
-		const response = await fetch("/api/devices");
-		const devices = await response.json();
-
-		const deviceFilter = document.getElementById("deviceFilter");
-		deviceFilter.innerHTML = '<option value="">All Devices</option>';
-
-		devices.forEach((device) => {
-			const option = document.createElement("option");
-			option.value = device;
-			option.textContent = device;
-			deviceFilter.appendChild(option);
-		});
-	} catch (error) {
-		console.error("Error loading devices:", error);
-		showAlert("Failed to load devices", "warning");
-	}
-}
-
-async function loadFirmwares() {
-	try {
-		const deviceType = document.getElementById("deviceFilter").value;
-		const url = deviceType ? `/api/firmwares?device=${encodeURIComponent(deviceType)}` : "/api/firmwares";
-
-		const response = await fetch(url);
-		const firmwares = await response.json();
-
-		if (response.ok) {
-			displayFirmwares(firmwares);
-		} else {
-			showAlert("Failed to load firmwares", "danger");
-		}
-	} catch (error) {
-		console.error("Error loading firmwares:", error);
-		showAlert("Error loading firmwares", "danger");
-	}
-}
-
-function displayFirmwares(firmwares) {
-	const container = document.getElementById("firmwaresList");
-
-	if (firmwares.length === 0) {
-		container.innerHTML = `
-            <div class="text-center py-5">
-                <i class="fas fa-inbox fa-3x text-muted mb-3"></i>
-                <h5>No firmwares available</h5>
-                <p class="text-muted">Upload your first firmware to get started.</p>
-            </div>
-        `;
-		return;
-	}
-
-	const grouped = firmwares.reduce((acc, firmware) => {
-		if (!acc[firmware.deviceType]) {
-			acc[firmware.deviceType] = [];
-		}
-		acc[firmware.deviceType].push(firmware);
-		return acc;
-	}, {});
-
-	// Sort each device type's firmwares by semver version (descending - newest first)
-	Object.keys(grouped).forEach((deviceType) => {
-		grouped[deviceType].sort((a, b) => compareSemver(b.version, a.version));
-	});
-
-	if (currentView === "list") {
-		displayFirmwaresListView(grouped);
-	} else {
-		displayFirmwaresGridView(grouped);
-	}
-}
-
-function displayFirmwaresGridView(grouped) {
-	const container = document.getElementById("firmwaresList");
-
-	container.innerHTML = Object.keys(grouped)
-		.sort() // Sort device types alphabetically
-		.map((deviceType) => {
-			const deviceFirmwares = grouped[deviceType];
-			return `
-                <div class="card mb-4">
-                    <div class="card-header bg-primary text-white">
-                        <div class="d-flex justify-content-between align-items-center">
-                            <h5 class="mb-0">
-                                <i class="fas fa-microchip me-2"></i>
-                                ${deviceType}
-                            </h5>
-                            <small>${deviceFirmwares.length} version${deviceFirmwares.length !== 1 ? "s" : ""}</small>
-                        </div>
-                    </div>
-                    <div class="card-body">
-                        <div class="row">
-                            ${deviceFirmwares.map((firmware) => createFirmwareCard(firmware)).join("")}
-                        </div>
-                    </div>
-                </div>
-            `;
-		})
-		.join("");
-}
-
-function displayFirmwaresListView(grouped) {
-	const container = document.getElementById("firmwaresList");
-
-	container.innerHTML = Object.keys(grouped)
-		.sort()
-		.map((deviceType) => {
-			const deviceFirmwares = grouped[deviceType];
-			return `
-                <div class="card mb-4">
-                    <div class="card-header bg-primary text-white">
-                        <div class="d-flex justify-content-between align-items-center">
-                            <h5 class="mb-0">
-                                <i class="fas fa-microchip me-2"></i>
-                                ${deviceType}
-                            </h5>
-                            <small>${deviceFirmwares.length} version${deviceFirmwares.length !== 1 ? "s" : ""}</small>
-                        </div>
-                    </div>
-                    <div class="card-body p-0">
-                        <div class="table-responsive">
-                            <table class="table table-hover mb-0">
-                                <thead class="table-light">
-                                    <tr>
-                                        <th>Version</th>
-                                        <th>Description</th>
-                                        <th>Size</th>
-                                        <th>Upload Date</th>
-                                        <th>SHA1</th>
-                                        <th>Actions</th>
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    ${deviceFirmwares.map((firmware) => createFirmwareRow(firmware)).join("")}
-                                </tbody>
-                            </table>
-                        </div>
-                    </div>
-                </div>
-            `;
-		})
-		.join("");
-}
-
-function createFirmwareRow(firmware) {
-	return `
-        <tr>
-            <td>
-                <span class="badge ${getVersionBadgeClass(firmware.version)} version-badge">${firmware.version}</span>
-            </td>
-            <td>
-                <div style="max-width: 300px; overflow: hidden; text-overflow: ellipsis; white-space: nowrap;" title="${firmware.description}">
-                    ${firmware.description}
-                </div>
-            </td>
-            <td>${formatFileSize(firmware.size)}</td>
-            <td>${formatDate(firmware.uploadDate)}</td>
-            <td>
-                <code class="sha1-hash" title="Click to copy" onclick="copyToClipboard('${firmware.sha1}', event)">
-                    ${firmware.sha1.substring(0, 8)}...
-                </code>
-            </td>
-            <td>
-                <div class="btn-group" role="group">
-                    <a href="/api/firmware/${firmware.id}/download" class="btn btn-outline-primary btn-sm px-3" title="Download">
-                        <i class="fas fa-download"></i>
-                    </a>
-
-                        <button class="btn btn-outline-warning btn-sm px-2" ${authToken ? `style="display: inline-block";` : `style="display: none";`} onclick="editFirmware('${
-		firmware.id
-	}')" title="Edit">
-                            <i class="fas fa-edit"></i>
-                        </button>
-                        <button class="btn btn-outline-danger btn-sm px-2" ${authToken ? `style="display: inline-block";` : `style="display: none";`} onclick="deleteFirmware('${
-		firmware.id
-	}')" title="Delete">
-                            <i class="fas fa-trash"></i>
-                        </button>
-                        <button class="btn btn-outline-info btn-sm px-3" onclick="viewFirmwareDetails('${firmware.id}')" title="View Details">
-                            <i class="fas fa-info"></i>
-                        </button>
-
-                </div>
-            </td>
-        </tr>
-    `;
-}
-
-function createFirmwareCard(firmware) {
-	return `
-        <div class="col-md-6 col-lg-4 mb-3">
-            <div class="card firmware-card h-100">
-                <div class="card-body">
-                    <div class="d-flex justify-content-between align-items-start mb-2">
-                        <span class="badge ${getVersionBadgeClass(firmware.version)} version-badge">${firmware.version}</span>
-                        <small class="text-muted">${formatDate(firmware.uploadDate)}</small>
-                    </div>
-                    <p class="card-text text-truncate">${firmware.description}</p>
-                    <div class="mb-2">
-                        <small class="text-muted d-block">
-                            <i class="fas fa-fingerprint me-1"></i>
-                            SHA1: <code class="sha1-hash" title="Click to copy" onclick="copyToClipboard('${firmware.sha1}', event)">${firmware.sha1.substring(
-		0,
-		8,
-	)}...</code>
-                        </small>
-                    </div>
-                    <div class="mt-auto">
-                        <div class="d-flex justify-content-between align-items-center mb-2">
-                            <small class="text-muted">${formatFileSize(firmware.size)}</small>
-                            <div class="btn-group" role="group">
-                                <a href="/api/firmware/${firmware.id}/download" class="btn btn-outline-primary btn-sm px-3" title="Download">
-                                    <i class="fas fa-download"></i>
-                                </a>
-
-                                    <button class="btn btn-outline-warning btn-sm px-2" ${authToken ? `style="display: inline-block";` : `style="display: none";`} onclick="editFirmware('${
-		firmware.id
-	}')" title="Edit">
-                                        <i class="fas fa-edit"></i>
-                                    </button>
-                                    <button class="btn btn-outline-danger btn-sm px-2"  ${authToken ? `style="display: inline-block"` : `style="display: none";`} onclick="deleteFirmware('${
-		firmware.id
-	}')" title="Delete">
-                                        <i class="fas fa-trash"></i>
-                                    </button>
-                                    <button class="btn btn-outline-info btn-sm px-3" onclick="viewFirmwareDetails('${firmware.id}')" title="View Details">
-                                    <i class="fas fa-info"></i>
-                                </button>
-
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    `;
-}
-
-async function viewFirmwareDetails(firmwareId) {
-	try {
-		const response = await fetch(`/api/firmware/${firmwareId}`);
-		const firmware = await response.json();
-
-		if (response.ok) {
-			showFirmwareModal(firmware);
-		} else {
-			showAlert("Failed to load firmware details", "danger");
-		}
-	} catch (error) {
-		console.error("Error loading firmware details:", error);
-		showAlert("Error loading firmware details", "danger");
-	}
-}
-
-function showFirmwareModal(firmware) {
-	// Create modal if it doesn't exist
-	let modal = document.getElementById("firmwareDetailsModal");
-	if (!modal) {
-		modal = document.createElement("div");
-		modal.innerHTML = `
-            <div class="modal fade" id="firmwareDetailsModal" tabindex="-1">
-                <div class="modal-dialog modal-lg">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <h5 class="modal-title">Firmware Details</h5>
-                            <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                        </div>
-                        <div class="modal-body" id="firmwareDetailsBody">
-                        </div>
-                        <div class="modal-footer">
-                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        `;
-		document.body.appendChild(modal);
-	}
-
-	const modalBody = document.getElementById("firmwareDetailsBody");
-	modalBody.innerHTML = `
-        <div class="row">
-            <div class="col-md-6">
-                <h6>Basic Information</h6>
-                <table class="table table-sm">
-                    <tr><td><strong>Device Type:</strong></td><td>${firmware.deviceType}</td></tr>
-                    <tr><td><strong>Version:</strong></td><td><span class="badge ${getVersionBadgeClass(firmware.version)}">${firmware.version}</span></td></tr>
-                    <tr><td><strong>File Size:</strong></td><td>${formatFileSize(firmware.size)}</td></tr>
-                    <tr><td><strong>Upload Date:</strong></td><td>${formatDate(firmware.uploadDate)}</td></tr>
-                    <tr><td><strong>Original Name:</strong></td><td>${firmware.originalName}</td></tr>
-                </table>
-            </div>
-            <div class="col-md-6">
-                <h6>Technical Details</h6>
-                <table class="table table-sm">
-                    <tr><td><strong>SHA1 Hash:</strong></td><td><code class="sha1-hash" title="Click to copy" onclick="copyToClipboard('${
-						firmware.sha1
-					}', event)">${firmware.sha1}</code></td></tr>
-                    <tr><td><strong>File ID:</strong></td><td><code>${firmware.fileId}</code></td></tr>
-                </table>
-            </div>
-        </div>
-        <div class="row mt-3">
-            <div class="col-12">
-                <h6>Description</h6>
-                <div class="border p-3 bg-light rounded">
-                    ${firmware.description}
-                </div>
-            </div>
-        </div>
-    `;
-
-	new bootstrap.Modal(document.getElementById("firmwareDetailsModal")).show();
-}
-
-async function editFirmware(firmwareId) {
-	try {
-		const response = await fetch(`/api/firmware/${firmwareId}`);
-		const firmware = await response.json();
-
-		if (response.ok) {
-			showEditFirmwareModal(firmware);
-		} else {
-			showAlert("Failed to load firmware details for editing", "danger");
-		}
-	} catch (error) {
-		console.error("Error loading firmware for edit:", error);
-		showAlert("Error loading firmware details", "danger");
-	}
-}
-
-function showEditFirmwareModal(firmware) {
-	// Create edit modal if it doesn't exist
-	let modal = document.getElementById("editFirmwareModal");
-	if (!modal) {
-		modal = document.createElement("div");
-		modal.innerHTML = `
-            <div class="modal fade" id="editFirmwareModal" tabindex="-1">
-                <div class="modal-dialog">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <h5 class="modal-title">Edit Firmware</h5>
-                            <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                        </div>
-                        <div class="modal-body">
-                            <form id="editFirmwareForm">
-                                <input type="hidden" id="editFirmwareId">
-                                <div class="mb-3">
-                                    <label for="editDeviceType" class="form-label">Device Type</label>
-                                    <input disabled type="text" class="form-control" id="editDeviceType" required>
-                                </div>
-                                <div class="mb-3">
-                                    <label for="editVersion" class="form-label">Version</label>
-                                    <input type="text" class="form-control version-badge" id="editVersion" required>
-                                </div>
-                                <div class="mb-3">
-                                    <label for="editDescription" class="form-label">Description</label>
-                                    <textarea class="form-control" id="editDescription" rows="3" required></textarea>
-                                </div>
-                            </form>
-                        </div>
-                        <div class="modal-footer">
-                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                            <button type="button" class="btn btn-primary" onclick="saveEditedFirmware()">Save Changes</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        `;
-		document.body.appendChild(modal);
-	}
-
-	// Populate form with current firmware data
-	document.getElementById("editFirmwareId").value = firmware.id;
-	document.getElementById("editDeviceType").value = firmware.deviceType;
-	document.getElementById("editVersion").value = firmware.version;
-	document.getElementById("editDescription").value = firmware.description;
-
-	new bootstrap.Modal(document.getElementById("editFirmwareModal")).show();
-}
-
-async function saveEditedFirmware() {
-	const firmwareId = document.getElementById("editFirmwareId").value;
-	const formData = {
-		deviceType: document.getElementById("editDeviceType").value,
-		version: document.getElementById("editVersion").value,
-		description: document.getElementById("editDescription").value,
-	};
-
-	try {
-		const response = await fetch(`/api/firmware/${firmwareId}`, {
-			method: "PUT",
-			headers: {
-				"Content-Type": "application/json",
-				Authorization: `Bearer ${authToken}`,
-			},
-			body: JSON.stringify(formData),
-		});
-
-		if (response.ok) {
-			showAlert("Firmware updated successfully", "success");
-			bootstrap.Modal.getInstance(document.getElementById("editFirmwareModal")).hide();
-			loadFirmwares();
-			loadDevices();
-			loadFirmwareStats();
-		} else {
-			const data = await response.json();
-			showAlert(data.error || "Failed to update firmware", "danger");
-		}
-	} catch (error) {
-		console.error("Error updating firmware:", error);
-		showAlert("Error updating firmware", "danger");
-	}
-}
-
-async function deleteFirmware(firmwareId) {
-	if (!confirm("Are you sure you want to delete this firmware? This action cannot be undone.")) {
-		return;
-	}
-
-	try {
-		const response = await fetch(`/api/firmware/${firmwareId}`, {
-			method: "DELETE",
-			headers: { Authorization: `Bearer ${authToken}` },
-		});
-
-		if (response.ok) {
-			showAlert("Firmware deleted successfully", "success");
-			loadFirmwares();
-			loadFirmwareStats();
-		} else {
-			const data = await response.json();
-			showAlert(data.error || "Failed to delete firmware", "danger");
-		}
-	} catch (error) {
-		console.error("Error deleting firmware:", error);
-		showAlert("Error deleting firmware", "danger");
-	}
-}
-
-async function uploadFirmware(e) {
-	e.preventDefault();
-
-	const formData = new FormData();
-	formData.append("deviceType", document.getElementById("deviceType").value);
-	formData.append("version", document.getElementById("version").value);
-	formData.append("description", document.getElementById("description").value);
-	formData.append("firmware", document.getElementById("firmwareFile").files[0]);
-
-	try {
-		const response = await fetch("/api/firmware/upload", {
-			method: "POST",
-			headers: { Authorization: `Bearer ${authToken}` },
-			body: formData,
-		});
-
-		const data = await response.json();
-
-		if (response.ok) {
-			showAlert("Firmware uploaded successfully!", "success");
-			document.getElementById("uploadForm").reset();
-			loadDevices();
-			loadFirmwares();
-			loadFirmwareStats();
-		} else {
-			showAlert(data.error || "Upload failed", "danger");
-		}
-	} catch (error) {
-		console.error("Upload error:", error);
-		showAlert("Upload failed - server error", "danger");
-	}
-}
-
-function copyToClipboard(text, event) {
-	event.stopPropagation();
-
-	if (navigator.clipboard && window.isSecureContext) {
-		navigator.clipboard
-			.writeText(text)
-			.then(() => {
-				showAlert("Hash copied to clipboard!", "success");
-			})
-			.catch(() => {
-				fallbackCopyToClipboard(text);
-			});
-	} else {
-		fallbackCopyToClipboard(text);
-	}
-}
-
-function fallbackCopyToClipboard(text) {
-	const textArea = document.createElement("textarea");
-	textArea.value = text;
-	textArea.style.position = "fixed";
-	textArea.style.left = "-999999px";
-	textArea.style.top = "-999999px";
-	document.body.appendChild(textArea);
-	textArea.focus();
-	textArea.select();
-
-	try {
-		document.execCommand("copy");
-		showAlert("Hash copied to clipboard!", "success");
-	} catch (err) {
-		showAlert("Failed to copy hash", "danger");
-	}
-
-	document.body.removeChild(textArea);
-}
-
-function showAlert(message, type) {
-	const toastContainer = document.getElementById("toastContainer");
-	const toastId = `toast_${Date.now()}`;
-	const toastDiv = document.createElement("div");
-	toastDiv.id = toastId;
-	toastDiv.className = `toast mb-2`;
-	toastDiv.setAttribute("role", "alert");
-	toastDiv.setAttribute("aria-live", "assertive");
-	toastDiv.setAttribute("aria-atomic", "true");
-
-	const colors = {
-		success: { bg: "#d1e7dd", border: "#badbcc", text: "#0f5132" },
-		danger: { bg: "#f8d7da", border: "#f5c2c7", text: "#842029" },
-		info: { bg: "#d1ecf1", border: "#bee5eb", text: "#055160" },
-		warning: { bg: "#fff3cd", border: "#ffecb5", text: "#664d03" },
-	};
-
-	const color = colors[type] || colors.info;
-
-	toastDiv.innerHTML = `
-        <div class="toast-body d-flex align-items-center" style="background-color: ${color.bg}; color: ${color.text}; border: 1px solid ${color.border}; border-radius: 0.375rem;">
-            <button type="button" class="btn-close me-3" onclick="removeToast('${toastId}')" style="margin: 0; padding: 0; width: 16px; height: 16px;"></button>
-            <div class="flex-grow-1">${message}</div>
-        </div>
-    `;
-
-	toastContainer.appendChild(toastDiv);
-
-	setTimeout(() => {
-		toastDiv.classList.add("show");
-	}, 100);
-
-	setTimeout(() => {
-		removeToast(toastId);
-	}, 5000);
-}
-
-function removeToast(toastId) {
-	const toast = document.getElementById(toastId);
-	if (toast) {
-		toast.classList.remove("show");
-		setTimeout(() => {
-			toast.remove();
-		}, 300);
-	}
-}
-
-function getVersionBadgeClass(version) {
-	const lowerVersion = version.toLowerCase();
-
-	if (lowerVersion.includes("alpha")) {
-		return "bg-danger";
-	} else if (lowerVersion.includes("beta")) {
-		return "bg-warning";
-	} else if (lowerVersion.includes("rc") || lowerVersion.includes("release-candidate")) {
-		return "bg-info";
-	} else if (lowerVersion.includes("dev") || lowerVersion.includes("snapshot")) {
-		return "bg-dark";
-	}
-	return "bg-success";
-}
-
-function compareSemver(a, b) {
-	const parseVersion = (version) => {
-		const cleaned = version.replace(/^v/, "");
-		return cleaned.split(".").map((part) => {
-			const [main, prerelease] = part.split("-");
-			const num = parseInt(main) || 0;
-			if (prerelease) {
-				return [num, prerelease];
-			}
-			return [num, null];
-		});
-	};
-
-	const versionA = parseVersion(a);
-	const versionB = parseVersion(b);
-	const maxLength = Math.max(versionA.length, versionB.length);
-
-	for (let i = 0; i < maxLength; i++) {
-		const partA = versionA[i] || [0, null];
-		const partB = versionB[i] || [0, null];
-
-		if (partA[0] !== partB[0]) {
-			return partA[0] - partB[0];
-		}
-
-		if (partA[1] !== partB[1]) {
-			if (partA[1] === null) return 1;
-			if (partB[1] === null) return -1;
-			return partA[1].localeCompare(partB[1]);
-		}
-	}
-
-	return 0;
-}
-
-function formatDate(dateString) {
-	return new Date(dateString).toLocaleDateString();
-}
-
-function formatFileSize(bytes) {
-	if (bytes === 0) return "0 Bytes";
-	const k = 1024;
-	const sizes = ["Bytes", "KB", "MB", "GB"];
-	const i = Math.floor(Math.log(bytes) / Math.log(k));
-	return `${parseFloat((bytes / Math.pow(k, i)).toFixed(2))} ${sizes[i]}`;
-}
